requirements.txt
fastapi==0.115.11
langchain==0.3.20
langchain_chroma==0.2.2
langchain_community==0.3.19
langchain_core==0.3.44
langchain_openai==0.3.8
langchain_text_splitters==0.3.6
loguru==0.7.3
pydantic==2.10.6
pydantic_settings==2.8.1
pymongo==4.11.2
pytest==8.3.5
python-dotenv==1.0.1
Requests==2.32.3
tqdm==4.66.2
tqdm==4.67.1
uvicorn==0.34.0
